Add some products to the docs page
Specifically, we had queues, cloud monitoring and auto scale in the left-hand side bar but not in the actual body of the page.  Likewise, I added Cloud Backup to the sidebar although it was in the main body of the page.

I also reordered the page to match the order of the sidebar.

A couple questions remain:

1. Should there be a Getting started guide for Cloud Backup?
2. What should the order be?
3. where do we get images from?

Thanks